rename block -> ban
