+++ packages/react/src/themes/teams/components/Icon/svg/ProcessedIcons/icons-block.tsx
@@ -17,5 +17,5 @@ export default {
     </svg>
   ),
   styles: {},
-  exportedAs: 'block',
+  exportedAs: 'ban',
 } as TeamsProcessedSvgIconSpec
